pdf_reading/requirements.txt
# Ferramentas de desenvolvimento
black
flake8
ipython
isort
jupyterlab
matplotlib
mkdocs
notebook
numpy
pandas
pip
pytest
python-dotenv
scikit-learn

# Bibliotecas para LLM e CrewAI
crewai>=0.1.0
langchain>=0.1.0
langchain-openai
openai>=1.0.0
anthropic
google-generativeai

# Processamento de PDFs
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
pdfplumber>=0.9.0

# Embedding e vetorização
chromadb>=0.4.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# Utilities
streamlit>=1.28.0
gradio>=3.50.0
python-multipart
requests>=2.31.0

-e .
